+++ ScreenCastingApp/ScreenCastingApp/ScreenGrabber.cs
@@ -5,6 +5,7 @@ using System.Drawing.Imaging;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
@@ -12,8 +13,38 @@ namespace ScreenCastingApp
 {
     class ScreenGrabber
     {
+        private readonly System.Timers.Timer timer;
 
-        public byte[] CaptureScreenshotAsBytes()
+        private byte[] currentImageData;
+
+        public ScreenGrabber()
+        {
+            timer = new System.Timers.Timer(150);
+            timer.Elapsed += TimerElapsed;
+        }
+
+        public void Start()
+        {
+            timer.Start();
+        }
+
+        public byte[] WaitForNextAvailableImage()
+        {
+            byte[] imageData = currentImageData;
+            while (imageData == null)
+            {
+                imageData = currentImageData;
+                System.Threading.Thread.Sleep(50);
+            }
+            return imageData;
+        }
+
+        private void TimerElapsed(object sender, System.Timers.ElapsedEventArgs e)
+        {
+            Interlocked.Exchange(ref currentImageData, CaptureScreenshotAsBytes());
+        }
+
+        private byte[] CaptureScreenshotAsBytes()
         {
             var offScreenImage = new Bitmap(Screen.PrimaryScreen.Bounds.Width,
                                             Screen.PrimaryScreen.Bounds.Height,
